Revert JNI calls to using dots instead of slashes
Bootclasspath classloader is not compatible with slashes which led to
HttpEngine crashing in very subtle manner. It happens when dexpreopt
fails which means that the bootclasspath classes are not cached so
we have to load them using JIT, the codepath for JIT loading only
accepts dots, while the codepath for dexpreopt accepts both. This is
why this crash was not observed in our CTS / MTS and most cases because
dexpreopt is the default behavior for all devices excluding -eng devices.

The longterm solution here is to fix the bootclasspath loader under JIT
to accept classes with slashes. This should be doable but requires
coordination with the ART team, so this CL is a hot-patch to make sure
that we dont release a broken version of Cronet.

Test: Manual test using -eng image.
Bug: 351821040
Change-Id: Ie1e8e977fcb749ad6f8465c93a95f68e81782526