Ignore some niche scenarios where Android.bp is not created
* Disable third_party/zstd:headers, this is because
Cronet is not using zstd. However, it bundles its header files
unconditionally due to b/361298517 so we never imported it but the
script is trying to create some "empty" targets there.

* Disable boringssl/fiat/fiat_license as boringssl is handled manually
for FIPS compliance. The autogenerated boringssl is never shipped,
just used for testing.

* Disable testing/buildbot:* as it contains a bunch of test filters
that are not used at all.

Bug: 358541596
Change-Id: Ia5476ea26120f5526d398f92932225990063dc53